+++ NEWS.md
@@ -2,6 +2,8 @@
 * Added `makeCaseWhenVariable()` that helps with many common recoding needs.
 * Added `alterArrayExpr()` which allows adding, removing, reordering and renaming subvariables
   in a derived Array.
+* You can now use the `keyring` package to store your credentials. See `?login()` for details 
+  (thanks @mainwaringb!)
 
 # crunch 1.27.1
 * Add several new `expressions` that let you create derived variables in more flexible ways than was previously possible.